user_service/requirements.txt
aiohttp==3.6.1
PyYAML==5.1.2
swagger-ui-bundle==0.0.5
aiohttp_jinja2==1.1.0
python-dateutil==2.8.0
asyncpg==0.19.0
connexion==2.4.0
